phy: synchronization improvements
Notes:
  - Don't send a sync right on startup.
  - Be more deterministic about re-starting as boss.
    - Time domain shifts can remain virtually synchronized.
    - Low risk of multi-boss scenario, often self-corrects.

This method of synchronization is an attempt to keep sync status
changes timed on the original sync pulses. There is a possibility
where two or more sync sources become unaware of each other because
they are synchronized and not seeing each others' syncs. In this
scenario, I have most commonly seen long-term sync without drift,
sometimes faster drift, and degraded performance during this time.
It seems that if it gets bad enough it always eventually
self-corrects. This situation is very rare at this point so I am
satisfied with the current strategy.

Signed-off-by: Phillip Sitbon <phillip.sitbon@intel.com>