phy: change rx off mode from idle to rx and set proper cs rssi thresh
Does not cause thrashing. Actually fixes a lot of problems because
it was not re-entering rx fast enough which mangled the packets.
Discovered while testing max rate mode and the change works very well.

Turns out the RSSI threshold was always wrong, in that it is relative
to AGC_GAIN_ADJUST.

Signed-off-by: Phillip Sitbon <phillip.sitbon@intel.com>