flsh: copy init data to ram and provide commit functionality
Now any data declared in .user can have an initial value, be
modified like normal data, and flushed to flash manually when
data is changed.

Next step may be to add a periodic (?) auto-flush function,
or an update function that takes the address and its new value
to combine a write-flush operation.

Signed-off-by: Phillip Sitbon <phillip.sitbon@intel.com>